+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.1.6.Pre.1
 - Allowed you to use old api system as a backup in case the current one is down or cannot be accessed
+- Add team death notifications that show when your teammates are killed or final killed
 
 ## Version 0.1.5
 - Allow customisable notification duration